backend-drf/api/urls.py
from django.urls import path
from accounts import views as userViews
from .views import StockPredictionAPIView
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path("register/", userViews.RegisterView.as_view()),
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('protected-view', userViews.ProtectedView.as_view()),

    #prediction
    path('predict/', StockPredictionAPIView.as_view(), name="stock_prediction")
]