requirement.txt
absl-py==1.4.0
asgiref==3.11.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.2
curl_cffi==0.13.0
cycler==0.12.1
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dm-tree==0.1.9
flatbuffers==25.12.19
fonttools==4.61.1
frozendict==2.4.7
gast==0.7.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
joblib==1.5.3
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
multitasking==0.0.12
namex==0.1.0
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
protobuf==6.33.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.5
tensorflow==2.20.0
tensorflow-model-optimization==0.8.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
Werkzeug==3.1.5
wrapt==1.14.2
yfinance==1.0